Move draft file detection to drafts.ts
Isolates detection logic away from the obsidian API
